Basic v0.2 interface structuring done. Prepare for implementation
gotcha: looks like only the top cmd 'cli' can initialize an object inside it's obj attribute. Some click implementation detail prevents objects to be registered to the context inside a sub command
when the obj attribute is then passed to a sub command, it looks like 'obj' == None. Probably best to change the implementation of the Age class a bit to change it's intent to merely a state class