The interface of the Age class is refactored slightly. No longer are the attributes of your age needed during object construction. This is done to facillitate to the click context behavior.
It is possible to set and re-set date and time, however a separate method for adding time info is provided. It only makes sense too add time if a date object was created first, so there is an internal check to enforce this.

the private methods _calculate and _convert are refactored too. _calculate used to modify the conversion register in place with the calculated value, tick only decided when to or not to call _convert as an extra step. However, it is a bit clearer to read the code, now that _calculate actually only calculates and returns the value up to the tick, who modifies the conversion register and reports back to the method caller.